+++ Core/Tools.Math.TravellingSalesman/LinKernighan/Edge.cs
@@ -26,7 +26,7 @@ namespace Tools.Math.TSP.LK
     {
         public int From { get; set; }
         public int To { get; set; }
-        public float Weight { get; set; }
+        public double Weight { get; set; }
 
         public override string ToString()
         {
